Service layer added

+++ WebFood/Controllers/RestaurantController.cs
@@ -5,15 +5,29 @@ using WebFood.Models.ViewModels;
 using WebFood.Utility;
 using WebPlanner.Models;
 using System.IO;
+using WebFood.Service.RestaurantService;
+using WebFood.Service.CategoryService;
+using WebFood.Service.TypeOfRestaurantService;
+using WebFood.Service.RestaurantTypeService;
+using WebFood.Service.UserService;
 
 namespace WebFood.Controllers
 {
     public class RestaurantController : Controller
     {
-        private readonly AppDbContext _db;
-        public RestaurantController(AppDbContext db)
+        private readonly IDaoRestaurant _daoRestaurant;
+        private readonly IDaoTypeOfRestaurant _daoTypeOfRestaurant;
+        private readonly IDaoRestaurantType _daoRestaurantType;
+        private readonly IDaoUser _daoUser;
+        public RestaurantController(IDaoRestaurant daoRestaurant,
+                                    IDaoTypeOfRestaurant daoTypeOfRestaurant,
+                                    IDaoRestaurantType daoRestaurantType,
+                                    IDaoUser daoUser)
         {
-            _db = db;
+            _daoRestaurant= daoRestaurant;
+            _daoTypeOfRestaurant = daoTypeOfRestaurant;
+            _daoRestaurantType = daoRestaurantType;
+            _daoUser= daoUser;
         }
         public IActionResult Restaurant(int restaurantId)
         {
@@ -28,17 +42,38 @@ namespace WebFood.Controllers
             var restaurant = GetRestaurant(restaurantId);
             AddRestaurantVM restaurantVM = new AddRestaurantVM();
             restaurantVM.Restaurant = restaurant;
-            restaurantVM.CategoryId = GetRestaurantCategory(restaurantId);
-            GetRestaurantCategories();
+            restaurantVM.CategoryId = GetRestaurantType(restaurantId);
+            GetTypesOfRestaurant();
             return View(restaurantVM);
         }
 
         [HttpPost]
         public IActionResult Edit(AddRestaurantVM restaurantVM)
         {
-            _db.Restaurants.Update(restaurantVM.Restaurant);
-            _db.SaveChanges();
-            ViewBag.Message = "Информация изменена";
+            var restaurant = restaurantVM.Restaurant;
+            if (restaurant.ManagerId == 0) restaurant.ManagerId = null;
+
+            if (ModelState.IsValid)
+            {
+
+                if (restaurant.ManagerId != null)
+                {
+                    if (_daoUser.GetAsync(Convert.ToInt32(restaurant.ManagerId)).Result != null)
+                    {
+                        _daoRestaurant.Update(restaurantVM.Restaurant);
+                        ViewBag.Message = "Информация изменена";
+                    }
+                    else
+                    {
+                        ModelState.AddModelError("Restaurant.ManagerId", "Пользователь с таким id не найден");
+                    }
+                }
+                else
+                {
+                    _daoRestaurant.Update(restaurantVM.Restaurant);
+                    ViewBag.Message = "Информация изменена";
+                }
+            }
             return View(restaurantVM);
         }
 
@@ -52,20 +87,23 @@ namespace WebFood.Controllers
         [HttpPost]
         public IActionResult ChangeImage(Restaurant restaurant, [FromForm(Name = "Imageurl")] IFormFile Imageurl)
         {
-            restaurant = GetRestaurant(restaurant.Id);
+            if (ModelState.IsValid)
+            {
+                restaurant = GetRestaurant(restaurant.Id);
 
-            string filePath = "wwwroot/"+restaurant.Imageurl;
+                string filePath = "wwwroot/"+restaurant.Imageurl;
 
 
-            if (System.IO.File.Exists(filePath)) {
-                System.IO.File.Delete(filePath);
+                if (System.IO.File.Exists(filePath)) {
+                    System.IO.File.Delete(filePath);
+                }
+
+                restaurant.Imageurl = GetImageUrl(Imageurl).Result.ToString();
+
+                _daoRestaurant.Update(restaurant);
+
+                ViewBag.Message = "Обложка изменена";
             }
-
-            restaurant.Imageurl = GetImageUrl(Imageurl).Result.ToString();
-
-            _db.Restaurants.Update(restaurant);
-            _db.SaveChanges();
-            ViewBag.Message = "Обложка изменена";
             return View(restaurant);
         }
 
@@ -84,15 +122,15 @@ namespace WebFood.Controllers
 
         private Restaurant GetRestaurant(int restaurantId)
         {
-            return _db.Restaurants.FirstOrDefault(r => r.Id == restaurantId);
+            return _daoRestaurant.GetAsync(restaurantId).Result;
         }
-        private int GetRestaurantCategory(int restaurantId)
+        private int GetRestaurantType(int restaurantId)
         {
-            return _db.RestaurantType.FirstOrDefault(rt => rt.RestaurantId == restaurantId).TypeId;
+            return _daoRestaurantType.GetAsync(restaurantId).Result.TypeId;
         }
-        private void GetRestaurantCategories()
+        private void GetTypesOfRestaurant()
         {
-            var categories = _db.Categories.ToList();
+            var categories = _daoTypeOfRestaurant.GetAllAsync().Result;
             ViewBag.RestaurantCategories = new SelectList(categories, "Id", "Name");
         }
     }
